... and upgrade and recompile deps

+++ test/test_LSIIR.py
@@ -280,9 +280,7 @@ def test_fitIIR_results_against_former_implementations(
     if inv:
         # Make sure there are non-zero frequency responses. Otherwise, fitting to
         # reciprocal of frequency response means dividing by zero.
-        assume(
-            not np.all(np.absolute(lsiir_base_params["H"]) < np.finfo(np.float64).eps)
-        )
+        assume(not np.all(np.absolute(lsiir_base_params["H"]) < np.finfo(float).eps))
 
     # Initialize parameters.
     fit_params = {
